Sensor: give every RNG buffer its own stream under a fixed seed
SetRandomSeed stored one value and nine call sites handed it straight to
init_cuda_rng. cuRAND separates generators within a buffer by its subsequence
parameter, and that parameter is already spent on the per-pixel index, so the
seed is the only thing telling one buffer's sequence from another's. One seed
at nine sites therefore made them one stream: two 64x64 buffers produced 4096
identical draws out of 4096, and a 640x360 buffer was a bit-exact prefix of a
1280x720 one. Any study that enabled reproducibility and used more than one
stochastic sensor got correlated noise between sensors, with nothing reported.
A single-sensor scene looks perfectly correct, which is why the existing suite
passed with the defect present: no test in it built two stochastic sensors.

Each buffer now derives its own seed from a stream key, bit-packed into 64 bits:

  [63:33] manager id   [32:13] sensor ordinal   [12:5] filter stream   [4:0] usage

The fields occupy disjoint bit ranges and each is range-checked against its
width, so distinct keys cannot collide. Adding the base seed cannot merge two
ids either, since x -> base + x is a bijection on the 64-bit ring for a fixed
base. Out-of-range fields throw rather than wrap, because wrapping would map two
sensors onto one stream, which is the defect.

Why each field exists:
  usage           separates purposes, via a closed RngUsage enum, one constant
                  per stochastic buffer in the module.
  sensor ordinal  separates instances of one purpose, minted in AddSensor.
  filter stream   separates two filters of the same purpose on one sensor, which
                  PushFilter permits since it applies no type check.
  manager id      separates managers that exist at the same time, whose ordinals
                  both restart at zero while the base seed is process-global. It
                  is the lowest free slot, released on destruction, NOT a
                  counter: a counter would give a rebuilt manager a different id
                  and break reproducibility across repeated construction, which
                  the byte-exact render tests rely on.

Identity is assigned in AddSensor rather than only in PushFilter, because
PushFilter is not the only route into a sensor's filter list. Seven sensor
classes fill the protected m_filters directly in their constructors, and
ChPhysCameraSensor puts a cuRAND-owning noise filter there. Every sensor passes
through AddSensor, so that is where the assignment cannot be bypassed.

GetRandomSeed() is deleted rather than deprecated, so a missed or future call
site is a compile error instead of a silent correlation. init_cuda_rng and its
kernel are widened to 64-bit to match curand_init and to carry the whole key.
The fixed-seed state is now mutex-guarded and read as a pair, so a concurrent
clear cannot be observed as "fixed, value zero". ChSensorManager is no longer
copyable, since it owns an identity slot its destructor releases.

Unchanged on purpose: sensor names are never generated or used as a key;
ChFilterPhysCameraNoise keeps its caller-supplied m_FPN_seed for the
fixed-pattern buffer and routes only shot noise through the new helper; with no
fixed seed every buffer still reads the clock afresh.

Tests: utest_SEN_rng_streams adds 22 cases needing no GPU, covering injectivity
over 10k ordinals by 20 purposes, all four fields varying together, out-of-range
rejection, duplicate same-purpose filters, filters added without PushFilter,
coexisting managers, reproducibility across rebuilds, slot reuse, the
unseeded path, a topology fuzzer over 40 randomized scenes, and a birthday-bound
check over 90k derived seeds. utest_SEN_camera_convergence adds render-level
cases: two cameras must differ under one fixed seed with and without noise
filters, the pair must stay reproducible across rebuilds, two physical cameras
must get independent shot-noise streams, and a buffer-level check that a 640x360
buffer is no longer a prefix of a 1280x720 one. That last one predicts its own
threshold, 1/(2s*sqrt(pi)) = 2.2% byte agreement for independent streams, and
measures 2.1752%, against 100% for a prefix.

The render-level cases fail against the pre-fix code with mean |diff| 0, two
cameras producing byte-identical frames, and pass after.

+++ src/chrono_sensor/cuda/curand_utils.cu
@@ -21,15 +21,20 @@
 namespace chrono {
 namespace sensor {
 
-__global__ void init_random_states(unsigned int seed, curandState_t* rng_states, unsigned int n_generators) {
-    int index = blockDim.x * blockIdx.x + threadIdx.x;
+// The seed is 64-bit because curand_init's seed parameter is, and because callers pack a stream
+// identity (manager, sensor, filter, purpose) into it. Narrowing to 32 bits here would truncate
+// the high fields of that identity and silently re-merge streams that the caller separated.
+__global__ void init_random_states(unsigned long long seed, curandState_t* rng_states, unsigned int n_generators) {
+    // unsigned throughout: the count cannot be negative, and mixing an int index with an unsigned
+    // bound would compare them after converting the index to unsigned.
+    unsigned int index = blockDim.x * blockIdx.x + threadIdx.x;
     if (index < n_generators)
         curand_init(seed, index, 0, &rng_states[index]);
 }
 
-void init_cuda_rng(unsigned int seed, curandState_t* rng_states, int n_generators) {
-    const int nThreads = 512;
-    int nBlocks = (n_generators + nThreads - 1) / nThreads;
+void init_cuda_rng(unsigned long long seed, curandState_t* rng_states, unsigned int n_generators) {
+    const unsigned int nThreads = 512;
+    unsigned int nBlocks = (n_generators + nThreads - 1) / nThreads;
     init_random_states<<<nBlocks, nThreads>>>(seed, rng_states, n_generators);
 }
